git-svn-id: svn://123.57.173.36/meijia-parent@2039 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-dao/src/main/java/com/simi/po/model/msg/Msg.java
@@ -9,7 +9,7 @@ public class Msg {
 
     private String gotoUrl;
 
-    private Short userType;
+    private Long userType;
 
     private Long sendTime;
 
@@ -61,11 +61,11 @@ public class Msg {
         this.gotoUrl = gotoUrl == null ? null : gotoUrl.trim();
     }
 
-    public Short getUserType() {
+    public Long getUserType() {
         return userType;
     }
 
-    public void setUserType(Short userType) {
+    public void setUserType(Long userType) {
         this.userType = userType;
     }
